+++ src/components/Date.js
@@ -2,6 +2,7 @@ import React, { Component } from "react";
 import styled from "styled-components/native";
 import PropTypes from "prop-types";
 import m from "moment";
+import _ from "lodash";
 
 const DateText = styled.Text`
   padding-top: 8;
@@ -37,9 +38,9 @@ class DateTime extends Component {
       }
 
       const hours = Math.floor(m.duration(m(date).diff(m())).asMinutes() / 60);
-      const minutes = `${Math.floor(
+      const minutes = _.padStart(`${Math.floor(
         ((m.duration(m(date).diff(m())).asMinutes() / 60) % 1) * 60
-      )}`.padStart(2, "0");
+      )}`, 2, '0');
       return `${hours}:${minutes}`;
     }
     return null;